+++ src/main/java/com/example/FilmBuffs/dto/request/DeleteFilmRequest.java
@@ -10,8 +10,7 @@ import lombok.*;
 @Builder
 public class DeleteFilmRequest {
 
-    private Long filmId;
-
+    private Integer filmId;
 
     public Film toFilm() {
         return Film.builder()
